seperated PushValue into component
